refactor: 결과 클래스의 인스턴스 변수 줄이기
