+++ src/main/java/bridge/service/dto/response/BridgeResponseDto.java
@@ -1,23 +1,16 @@
 package bridge.service.dto.response;
 
+import bridge.domain.BridgeStatus;
 import bridge.domain.Result;
 
-import java.util.List;
-
 public class BridgeResponseDto {
-    private final List<String> upBlocks;
-    private final List<String> downBlocks;
+    private final BridgeStatus bridgeStatus;
 
     public BridgeResponseDto(Result result) {
-        this.upBlocks = result.getUpBlocks();
-        this.downBlocks = result.getDownBlocks();
+        this.bridgeStatus = result.getBridgeStatus();
     }
 
-    public List<String> getUpBlocks() {
-        return this.upBlocks;
-    }
-
-    public List<String> getDownBlocks() {
-        return this.downBlocks;
+    public BridgeStatus getBridgeStatus() {
+        return this.bridgeStatus;
     }
 }
