+++ src/main/java/bridge/service/dto/response/GameResultResponseDto.java
@@ -1,29 +1,24 @@
 package bridge.service.dto.response;
 
+import bridge.domain.BridgeStatus;
 import bridge.domain.Player;
 import bridge.domain.Result;
 
 import java.util.List;
 
 public class GameResultResponseDto {
-    private final List<String> upBlocks;
-    private final List<String> downBlocks;
+    private final BridgeStatus bridgeStatus;
     private final String gameSuccessOrFail;
     private final int totalTryNumber;
 
     public GameResultResponseDto(Result result, Player player) {
-        this.upBlocks = result.getUpBlocks();
-        this.downBlocks = result.getDownBlocks();
+        this.bridgeStatus = result.getBridgeStatus();
         this.gameSuccessOrFail = result.getGameOver().getName();
         this.totalTryNumber = player.getTotalTryNumber();
     }
 
-    public List<String> getUpBlocks() {
-        return upBlocks;
-    }
-
-    public List<String> getDownBlocks() {
-        return downBlocks;
+    public BridgeStatus getBridgeStatus() {
+        return bridgeStatus;
     }
 
     public String getGameSuccessOrFail() {
